requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1
torchvision==0.13.1
tensorboard==2.10.1
numpy==1.23.3
scikit-learn==1.1.3
matplotlib==3.6.1
wandb==0.13.4
pyhocon==0.3.59
pytest==7.2.0
protobuf==3.19.6
imath~=0.0.2
OpenEXR~=1.3.8
Pillow~=9.2.0
torchmetrics~=0.11.0